Update referenced NuGet packages.
Add "not" operator and allow AND/OR as operators.
Fix typecast error in Expression.cs.
Allow dash in words.
Expand multi word concepts such as "a bird" such that it can be matched by the words "a" and "bird".
Make sure output is UTF8 compatible - allow emojies in output.

+++ ZimmerBot.Core/Expressions/Expression.cs
@@ -26,7 +26,7 @@ namespace ZimmerBot.Core.Expressions
       else if (a is int)
         b = (int)a != 0;
       else if (a is double)
-        b = (int)a != 0d;
+        b = (double)a != 0d;
       else if (a is string)
         b = !string.IsNullOrEmpty((string)a);
       else if (a == null)
